@@ -1,6 +1,6 @@
 {
   "name": "@sentry/wizard",
-  "version": "3.34.4",
+  "version": "4.0.2",
   "homepage": "https://github.com/getsentry/sentry-wizard",
   "repository": "https://github.com/getsentry/sentry-wizard",
   "description": "Sentry wizard helping you to configure your project",
@@ -22,19 +22,19 @@
   "typescript": {
     "definition": "dist/index.d.ts"
   },
+  "files": [
+    "dist"
+  ],
   "dependencies": {
     "@clack/core": "^0.3.4",
     "@clack/prompts": "0.7.0",
-    "@sentry/cli": "^1.77.3",
     "@sentry/node": "^7.119.2",
     "axios": "1.7.4",
     "chalk": "^2.4.1",
-    "glob": "^8.1.0",
+    "glob": "9.3.5",
     "inquirer": "^6.2.0",
-    "lodash": "^4.17.15",
     "magicast": "^0.2.10",
     "opn": "^5.4.0",
-    "r2": "^2.0.1",
     "read-env": "^1.3.0",
     "recast": "^0.23.3",
     "semver": "^7.5.3",
@@ -43,13 +43,15 @@
     "yargs": "^16.2.0"
   },
   "devDependencies": {
+    "@babel/types": "~7.21.4",
     "@sentry-internal/eslint-config-sdk": "^7.48.0",
     "@types/chai": "^4.3.17",
     "@types/glob": "^7.2.0",
     "@types/inquirer": "^0.0.43",
     "@types/jest": "^29.5.0",
     "@types/lodash": "^4.14.144",
-    "@types/node": "^10.11.0",
+    "@types/node": "^18.19.76",
+    "@types/opn": "5.1.0",
     "@types/rimraf": "^3.0.2",
     "@types/semver": "^7.3.7",
     "@types/yargs": "^16.0.9",
@@ -59,6 +61,7 @@
     "eslint": "^8.18.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-jest": "^25.3.0",
+    "fossilize": "^0.3.1",
     "jest": "^29.5.0",
     "prettier": "^2.8.7",
     "rimraf": "^3.0.2",
@@ -67,8 +70,11 @@
     "tsx": "^3.14.0",
     "typescript": "^5.0.4"
   },
+  "resolutions": {
+    "simple-plist": "1.4.0-0"
+  },
   "engines": {
-    "node": ">=14.18.0",
+    "node": ">=18.20.6",
     "npm": ">=3.10.7",
     "yarn": ">=1.0.2"
   },
@@ -84,8 +90,9 @@
     "fix": "yarn fix:eslint && yarn fix:prettier",
     "fix:prettier": "prettier --write \"{lib,src,test}/**/*.ts\"",
     "fix:eslint": "eslint . --format stylish --fix",
-    "test": "yarn build && jest",
+    "test": "jest",
     "test:e2e": "yarn build && ./e2e-tests/run.sh",
+    "test:e2e:bin": "chmod +x ./dist-bin/* && SENTRY_WIZARD_E2E_TEST_BIN=1 ./e2e-tests/run.sh",
     "try": "ts-node bin.ts",
     "try:uninstall": "ts-node bin.ts --uninstall",
     "test:watch": "jest --watch"
@@ -107,7 +114,7 @@
       "json"
     ],
     "modulePathIgnorePatterns": [
-      "<rootDir>/dist/__mocks__"
+      "<rootDir>/dist/"
     ],
     "testPathIgnorePatterns": [
       "/dist/",
@@ -121,7 +128,8 @@
   "author": "Sentry",
   "license": "MIT",
   "volta": {
-    "node": "14.18.3",
+    "node": "18.20.6",
     "yarn": "1.22.19"
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
