@@ -1,4 +1,4 @@
-const readEnv = require('read-env').default;
+import readEnv from 'read-env';
 
 // TODO: move to src/utils (+tests)
 export function readEnvironment(): Record<string, unknown> {
