@@ -1,8 +1,22 @@
 import type { Answers } from 'inquirer';
 import { prompt } from 'inquirer';
-import * as _ from 'lodash';
 
 import { BaseStep } from './BaseStep';
+
+type Project = {
+  id: string;
+  slug: string;
+  organization?: {
+    name: string;
+    slug: string;
+  };
+  keys: {
+    dsn: {
+      public: string;
+      private: string;
+    };
+  }[];
+};
 
 function sleep(n: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, n));
@@ -12,23 +26,22 @@ export class SentryProjectSelector extends BaseStep {
   public async emit(answers: Answers): Promise<any> {
     this.debug(answers);
 
-    if (!_.has(answers, 'wizard')) {
+    if (!answers.wizard) {
       // we skip this completely because the wizard wasn't running
       return {};
     }
 
-    if (
-      _.has(answers, 'wizard.projects') &&
-      answers.wizard.projects.length === 0
-    ) {
+    if (!answers.wizard.projects || answers.wizard.projects.length === 0) {
       throw new Error(
         'No Projects found. Please create a new Project in Sentry and try again.',
       );
     }
 
-    let selectedProject = null;
+    let selectedProject: { selectedProject: Project } | null = null;
     if (answers.wizard.projects.length === 1) {
-      selectedProject = { selectedProject: answers.wizard.projects[0] };
+      selectedProject = {
+        selectedProject: answers.wizard.projects[0] as Project,
+      };
       // the wizard CLI closes too quickly when we skip the prompt
       // as it will cause the UI to be stuck saying Waiting for wizard to connect
       await sleep(1000);
@@ -47,34 +60,22 @@ export class SentryProjectSelector extends BaseStep {
         },
       ]);
     }
-
+    const dsn = selectedProject?.selectedProject.keys[0]?.dsn ?? {
+      public: null,
+      private: null,
+    };
     return {
       config: {
         auth: {
-          token: _.get(answers, 'wizard.apiKeys.token', null),
+          token: answers.wizard.apiKeys?.token ?? null,
         },
-        dsn: {
-          public: _.get(
-            selectedProject,
-            'selectedProject.keys.0.dsn.public',
-            null,
-          ),
-          secret: _.get(
-            selectedProject,
-            'selectedProject.keys.0.dsn.secret',
-            null,
-          ),
-        },
+        dsn,
         organization: {
-          slug: _.get(
-            selectedProject,
-            'selectedProject.organization.slug',
-            null,
-          ),
+          slug: selectedProject?.selectedProject.organization?.slug ?? null,
         },
         project: {
-          id: _.get(selectedProject, 'selectedProject.id', null),
-          slug: _.get(selectedProject, 'selectedProject.slug', null),
+          id: selectedProject?.selectedProject.id ?? null,
+          slug: selectedProject?.selectedProject.slug ?? null,
         },
       },
     };
