@@ -28,7 +28,7 @@ const sentryWebpackPluginOptions = {
 
   silent: true, // Suppresses all logs
   // For all available options, see:
-  // https://github.com/getsentry/sentry-webpack-plugin#options.
+  // https://www.npmjs.com/package/@sentry/webpack-plugin#options
 };
 
 // Make sure adding Sentry options is the last code to run before exporting, to
