@@ -98,6 +98,6 @@ Sentry.init({
   dsn: '${dsn}',
 
   // uncomment the line below to enable Spotlight (https://spotlightjs.com)
-  // enableSpotlight: __DEV__,
+  // spotlight: __DEV__,
 });`;
 }
