@@ -7,9 +7,9 @@ import {
   abortIfCancelled,
   addDotEnvSentryBuildPluginFile,
 } from '../../utils/clack-utils';
-import { WizardOptions } from '../../utils/types';
+import type { WizardOptions } from '../../utils/types';
 
-import { SourceMapUploadToolConfigurationOptions } from './types';
+import type { SourceMapUploadToolConfigurationOptions } from './types';
 
 import * as Sentry from '@sentry/node';
 
@@ -33,9 +33,6 @@ const getCodeSnippet = (options: SourceMapUploadToolConfigurationOptions) =>
   ${chalk.greenBright(`const sentryOptions = {
     // Upload additional client files (increases upload size)
     widenClientFileUpload: true,
-
-    // Hides source maps from generated client bundles
-    hideSourceMaps: true,
   };`)}
 
   ${chalk.greenBright(`module.exports = withSentryConfig(
