@@ -1,17 +1,18 @@
 import {
-  defaultStackParser,
   Hub,
   Integrations,
+  NodeClient,
+  type Span,
+  defaultStackParser,
+  flush,
   makeMain,
   makeNodeTransport,
-  NodeClient,
   runWithAsyncContext,
   setTag,
   startSpan,
-  flush,
 } from '@sentry/node';
-import packageJson from '../package.json';
-import { WizardOptions } from './utils/types';
+import type { WizardOptions } from './utils/types';
+import { WIZARD_VERSION } from './version';
 
 export async function withTelemetry<F>(
   options: {
@@ -76,7 +77,7 @@ function createSentryInstance(enabled: boolean, integration: string) {
     tracesSampleRate: 1,
     sampleRate: 1,
 
-    release: packageJson.version,
+    release: WIZARD_VERSION,
     integrations: [new Integrations.Http()],
     tracePropagationTargets: [/^https:\/\/sentry.io\//],
 
@@ -110,9 +111,12 @@ function createSentryInstance(enabled: boolean, integration: string) {
   return { sentryHub: hub, sentryClient: client };
 }
 
-export function traceStep<T>(step: string, callback: () => T): T {
+export function traceStep<T>(
+  step: string,
+  callback: (span: Span | undefined) => T,
+): T {
   updateProgress(step);
-  return startSpan({ name: step, op: 'wizard.step' }, () => callback());
+  return startSpan({ name: step, op: 'wizard.step' }, (span) => callback(span));
 }
 
 export function updateProgress(step: string) {
