@@ -1,39 +1,47 @@
 import {
   askForToolConfigPath,
+  askForWizardLogin,
   createNewConfigFile,
+  installPackage,
 } from '../../src/utils/clack-utils';
 
-import * as fs from 'fs';
+import * as fs from 'node:fs';
+import * as ChildProcess from 'node:child_process';
+import type { PackageManager } from '../../src/utils/package-manager';
 
-type ClackMock = {
-  confirm: jest.Mock;
-  text: jest.Mock;
-  isCancel: jest.Mock;
-  cancel: jest.Mock;
+import axios from 'axios';
+
+// @ts-ignore - clack is ESM and TS complains about that. It works though
+import * as clack from '@clack/prompts';
+
+jest.mock('node:child_process', () => ({
+  __esModule: true,
+  ...jest.requireActual('node:child_process'),
+}));
+
+jest.mock('@clack/prompts', () => ({
   log: {
-    info: jest.Mock;
-    success: jest.Mock;
-    warn: jest.Mock;
-  };
-};
+    info: jest.fn(),
+    success: jest.fn(),
+    warn: jest.fn(),
+    error: jest.fn(),
+  },
+  outro: jest.fn(),
+  text: jest.fn(),
+  confirm: jest.fn(),
+  cancel: jest.fn(),
+  // passthrough for abortIfCancelled
+  isCancel: jest.fn().mockReturnValue(false),
+  spinner: jest
+    .fn()
+    .mockImplementation(() => ({ start: jest.fn(), stop: jest.fn() })),
+}));
+const clackMock = clack as jest.Mocked<typeof clack>;
 
-let clackMock: ClackMock;
+jest.mock('axios');
+const mockedAxios = axios as jest.Mocked<typeof axios>;
 
-jest.mock('@clack/prompts', () => {
-  clackMock = {
-    log: {
-      info: jest.fn(),
-      success: jest.fn(),
-      warn: jest.fn(),
-    },
-    text: jest.fn(),
-    confirm: jest.fn(),
-    cancel: jest.fn(),
-    // passthrough for abortIfCancelled
-    isCancel: jest.fn().mockReturnValue(false),
-  };
-  return clackMock;
-});
+jest.mock('opn', () => jest.fn(() => Promise.resolve({ on: jest.fn() })));
 
 function mockUserResponse(fn: jest.Mock, response: any) {
   fn.mockReturnValueOnce(response);
@@ -45,11 +53,11 @@ describe('askForToolConfigPath', () => {
   });
 
   it('returns undefined if users have no config file', async () => {
-    mockUserResponse(clackMock.confirm, false);
+    mockUserResponse(clack.confirm as jest.Mock, false);
 
     const result = await askForToolConfigPath('Webpack', 'webpack.config.js');
 
-    expect(clackMock.confirm).toHaveBeenCalledWith(
+    expect(clack.confirm).toHaveBeenCalledWith(
       expect.objectContaining({
         message: expect.stringContaining('have a Webpack config file'),
       }),
@@ -59,18 +67,18 @@ describe('askForToolConfigPath', () => {
   });
 
   it('returns the path if users have a config file and the entered path is valid', async () => {
-    mockUserResponse(clackMock.confirm, true);
-    mockUserResponse(clackMock.text, 'my.webpack.config.js');
+    mockUserResponse(clack.confirm as jest.Mock, true);
+    mockUserResponse(clack.text as jest.Mock, 'my.webpack.config.js');
 
     const result = await askForToolConfigPath('Webpack', 'webpack.config.js');
 
-    expect(clackMock.confirm).toHaveBeenCalledWith(
+    expect(clack.confirm).toHaveBeenCalledWith(
       expect.objectContaining({
         message: expect.stringContaining('have a Webpack config file'),
       }),
     );
 
-    expect(clackMock.text).toHaveBeenCalledWith(
+    expect(clack.text).toHaveBeenCalledWith(
       expect.objectContaining({
         message: expect.stringContaining(
           'enter the path to your Webpack config file',
@@ -92,8 +100,8 @@ describe('createNewConfigFile', () => {
       .spyOn(fs.promises, 'writeFile')
       .mockImplementation(jest.fn());
 
-    const filename = '/weboack.config.js';
-    const code = `module.exports = {/*config...*/}`;
+    const filename = '/webpack.config.js';
+    const code = 'module.exports = {/*config...*/}';
 
     const result = await createNewConfigFile(filename, code);
 
@@ -104,14 +112,14 @@ describe('createNewConfigFile', () => {
   it('logs more information if provided as an argument', async () => {
     jest.spyOn(fs.promises, 'writeFile').mockImplementation(jest.fn());
 
-    const filename = '/weboack.config.js';
-    const code = `module.exports = {/*config...*/}`;
+    const filename = '/webpack.config.js';
+    const code = 'module.exports = {/*config...*/}';
     const moreInfo = 'More information...';
 
     await createNewConfigFile(filename, code, moreInfo);
 
-    expect(clackMock.log.info).toHaveBeenCalledTimes(1);
-    expect(clackMock.log.info).toHaveBeenCalledWith(
+    expect(clack.log.info).toHaveBeenCalledTimes(1);
+    expect(clack.log.info).toHaveBeenCalledWith(
       expect.stringContaining(moreInfo),
     );
   });
@@ -122,13 +130,13 @@ describe('createNewConfigFile', () => {
       .mockImplementation(() => Promise.reject(new Error('Could not write')));
 
     const filename = '/webpack.config.js';
-    const code = `module.exports = {/*config...*/}`;
+    const code = 'module.exports = {/*config...*/}';
 
     const result = await createNewConfigFile(filename, code);
 
     expect(result).toBe(false);
     expect(writeFileSpy).toHaveBeenCalledWith(filename, code);
-    expect(clackMock.log.warn).toHaveBeenCalledTimes(1);
+    expect(clack.log.warn).toHaveBeenCalledTimes(1);
   });
 
   it('returns false if the passed path is not absolute', async () => {
@@ -138,5 +146,144 @@ describe('createNewConfigFile', () => {
     );
 
     expect(result).toBe(false);
+  });
+});
+
+describe('installPackage', () => {
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('force-installs a package if the forceInstall flag is set', async () => {
+    const packageManagerMock: PackageManager = {
+      name: 'npm',
+      label: 'NPM',
+      installCommand: 'npm install',
+      buildCommand: 'npm run build',
+      runScriptCommand: 'npm run',
+      flags: '',
+      forceInstallFlag: '--force',
+      detect: jest.fn(),
+      addOverride: jest.fn(),
+    };
+
+    const execSpy = jest
+      .spyOn(ChildProcess, 'exec')
+      // @ts-expect-error - don't care about the return value
+      .mockImplementationOnce((cmd, cb) => {
+        if (cb) {
+          // @ts-expect-error - don't care about the options value
+          cb(null, '', '');
+        }
+      });
+
+    await installPackage({
+      alreadyInstalled: false,
+      packageName: '@sentry/sveltekit',
+      packageNameDisplayLabel: '@sentry/sveltekit',
+      forceInstall: true,
+      askBeforeUpdating: false,
+      packageManager: packageManagerMock,
+    });
+
+    expect(execSpy).toHaveBeenCalledWith(
+      'npm install @sentry/sveltekit  --force',
+      expect.any(Function),
+    );
+  });
+
+  it.each([false, undefined])(
+    "doesn't force-install a package if the forceInstall flag is %s",
+    async (flag) => {
+      const packageManagerMock: PackageManager = {
+        name: 'npm',
+        label: 'NPM',
+        installCommand: 'npm install',
+        buildCommand: 'npm run build',
+        runScriptCommand: 'npm run',
+        flags: '',
+        forceInstallFlag: '--force',
+        detect: jest.fn(),
+        addOverride: jest.fn(),
+      };
+
+      const execSpy = jest
+        .spyOn(ChildProcess, 'exec')
+        // @ts-expect-error - don't care about the return value
+        .mockImplementationOnce((cmd, cb) => {
+          if (cb) {
+            // @ts-expect-error - don't care about the options value
+            cb(null, '', '');
+          }
+        });
+
+      await installPackage({
+        alreadyInstalled: false,
+        packageName: '@sentry/sveltekit',
+        packageNameDisplayLabel: '@sentry/sveltekit',
+        forceInstall: flag,
+        askBeforeUpdating: false,
+        packageManager: packageManagerMock,
+      });
+
+      expect(execSpy).toHaveBeenCalledWith(
+        'npm install @sentry/sveltekit  ',
+        expect.any(Function),
+      );
+    },
+  );
+});
+
+describe('askForWizardLogin', () => {
+  // mock axios
+  afterEach(() => {
+    // clackMock.confirm.mockClear();
+    // mockUserResponse(clack.confirm as jest.Mock, undefined);
+  });
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+    mockedAxios.get.mockClear();
+    clackMock.confirm.mockClear();
+    clackMock.confirm.mockReset();
+    mockUserResponse(clack.confirm as jest.Mock, undefined);
+  });
+
+  it('asks if a user already has a Sentry account by default', async () => {
+    mockUserResponse(clack.confirm as jest.Mock, Promise.resolve(true));
+
+    // Provide the data object to be returned
+    mockedAxios.get.mockResolvedValue({
+      data: {
+        hash: 'mockedHash',
+      },
+    });
+
+    await askForWizardLogin({ url: 'https://santry.io/' });
+
+    expect(clack.confirm).toHaveBeenCalledWith(
+      expect.objectContaining({
+        message: expect.stringContaining('already have a Sentry account'),
+      }),
+    );
+  });
+
+  it('skips asking for if a user already has a Sentry account if org and project are pre-selected', async () => {
+    mockUserResponse(clackMock.confirm as jest.Mock, Promise.resolve(true));
+
+    // Provide the data object to be returned
+    mockedAxios.get.mockResolvedValue({
+      data: {
+        hash: 'mockedHash',
+      },
+    });
+
+    await askForWizardLogin({
+      url: 'https://santry.io/',
+      orgSlug: 'my-org',
+      projectSlug: 'my-project',
+    });
+
+    expect(clack.confirm).not.toHaveBeenCalled();
   });
 });
